Crash Type and Filler Dummies
Added crash type dummies as well as a function to create zero-value dummy columns for the values of manner collision and crash type that exist but don't occur in the data. Also created dummies that group important manners. Created a translation dictionary between CMF crash attributes and binary columns we have.