Added NotIntersection and Zero Cost CM Feature
Added an option to the Crash Type Translation Dictionary for Non-intersection crash types for CMFs. The field this will reference will be added by Max. Also added a catch for if the estimated cost of a countermeasure was zero to automatically set the BC Ratio to zero, avoiding a divide-by-zero error.